requirements.txt
pillow==12.*
streamlit
firebase-admin
pandas
python-dotenv
google-cloud-storage
plotly
streamlit-autorefresh